+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - HIndent now assumes no extensions are enabled by default ([#728]).
 - All modules except for `HIndent` are now private, and the minimum necessary definitions are exposed from the module ([#729]).
 - HIndent now prints all `do` expressions in a unified style  ([#739]).
-- HIndent now formats operators based on their fixities ([#741]).
+- HIndent now formats operators based on their fixities ([#741], [#742]).
 
 ### Fixed
 
@@ -351,6 +351,7 @@ This version is accidentally pushlished, and is the same as 5.3.3.
 [@uhbif19]: https://github.com/uhbif19
 [@toku-sa-n]: https://github.com/toku-sa-n
 
+[#742]: https://github.com/mihaimaruseac/hindent/pull/742
 [#741]: https://github.com/mihaimaruseac/hindent/pull/741
 [#739]: https://github.com/mihaimaruseac/hindent/pull/739
 [#731]: https://github.com/mihaimaruseac/hindent/pull/731